+++ packages/table/src/plugins/sticky/index.ts
@@ -1,1 +1,3 @@
 export * from './useSticky'
+export * from './useStickyColumn'
+export * from './StickyColumn'