+++ packages/table/src/plugins/sticky/state.ts
@@ -1,5 +1,4 @@
 import type { ColumnId, RowId } from '@grid-table/basic'
-import { columnIdShowListAtom, rowIdShowListAtom } from '@grid-table/basic'
 import { atom } from '@einfach/react'
 import type { AtomEntity } from '@einfach/react'
 import type { PositionType, StickyType } from './type'
@@ -13,36 +12,10 @@ export const stickyOptions = atom<{
   rowFixed?: boolean
 }>({})
 
-export const stickyLeftOptionAtom = atom<ColumnId[]>([])
-export const stickyRightOptionAtom = atom<ColumnId[]>([])
-export const stickyTopOptionAtom = atom<ColumnId[]>([])
-export const stickyBottomOptionAtom = atom<ColumnId[]>([])
-
-export const stickyLeftAtom = atom((getter) => {
-  const showIds = new Set(getter(columnIdShowListAtom))
-  return getter(stickyLeftOptionAtom).filter((id) => {
-    return showIds.has(id)
-  })
-})
-
-export const stickyRightAtom = atom((getter) => {
-  const showIds = new Set(getter(columnIdShowListAtom))
-  return getter(stickyRightOptionAtom).filter((id) => {
-    return showIds.has(id)
-  })
-})
-export const stickyTopAtom = atom((getter) => {
-  const showIds = new Set(getter(rowIdShowListAtom))
-  return getter(stickyTopOptionAtom).filter((id) => {
-    return showIds.has(id)
-  })
-})
-export const stickyBottomAtom = atom((getter) => {
-  const showIds = new Set(getter(rowIdShowListAtom))
-  return getter(stickyBottomOptionAtom).filter((id) => {
-    return showIds.has(id)
-  })
-})
+export const stickyLeftAtom = atom<ColumnId[]>([])
+export const stickyRightAtom = atom<ColumnId[]>([])
+export const stickyTopAtom = atom<ColumnId[]>([])
+export const stickyBottomAtom = atom<ColumnId[]>([])
 
 export const StickyConfig: Record<
   StickyType,